fixed DRY comment and updated code
